+++ src/test/java/net/kiel/dnd/service/CharacterServiceTest.java
@@ -35,7 +35,7 @@ public class CharacterServiceTest {
     
     @Test
     public void testGet() {
-        final Integer characterId = 1;
+        final Long characterId = 1L;
         Character character = characterService.findById(characterId);
         
         assertNotNull(character);
